Move file string extract function
