+++ SGGL-Diablo-Knowledge-Library/src/game_version.c
@@ -36,13 +36,8 @@
 #include "diablo/diablo_game_version.h"
 #include "diablo_ii/diablo_ii_game_version.h"
 #include "helper/error_handling.h"
+#include "helper/file_info.h"
 #include "helper/game_version_finder.h"
-
-/* Struct taken from Microsoft's example. */
-struct LANGANDCODEPAGE {
-  WORD wLanguage;
-  WORD wCodePage;
-};
 
 /*
 * The order of the entries should be in lexicographical order, due to
@@ -58,126 +53,6 @@ find_version_func_table[] = {
 
 static enum GameVersion running_game_version = VERSION_UNKNOWN;
 static wchar_t* running_product_name;
-
-static wchar_t* ExtractFileStringValue(
-    const wchar_t* game_path,
-    const wchar_t* string_name,
-    size_t string_name_len
-) {
-  DWORD ignored;
-
-  void* file_version_info;
-  DWORD file_version_info_size;
-
-  BOOL is_get_file_version_info_success;
-  BOOL is_ver_query_value_success;
-
-  struct LANGANDCODEPAGE* lang_buffer;
-  UINT lang_buffer_size;
-
-  wchar_t* temp_file_string_value;
-
-  wchar_t* file_string_value;
-  UINT file_string_value_size;
-
-  wchar_t* file_string_sub_block;
-  size_t file_string_sub_block_capacity;
-
-  int swprintf_result;
-
-  /* Check version size. */
-  file_version_info_size = GetFileVersionInfoSizeW(
-      game_path,
-      &ignored
-  );
-
-  if (file_version_info_size == 0) {
-    ExitOnWindowsFunctionFailureWithLastError(
-        L"GetFileVersionInfoSizeW",
-        GetLastError()
-    );
-  }
-
-  /* Get the file version info.*/
-  file_version_info = malloc(file_version_info_size);
-
-  if (file_version_info == NULL) {
-    ExitOnAllocationFailure();
-  }
-
-  is_get_file_version_info_success = GetFileVersionInfoW(
-      game_path,
-      ignored,
-      file_version_info_size,
-      file_version_info
-  );
-
-  if (!is_get_file_version_info_success) {
-    ExitOnWindowsFunctionFailureWithLastError(
-        L"GetFileVersionInfoW",
-        GetLastError()
-    );
-  }
-
-  /* Gather all of the information into the specified buffer. */
-  is_ver_query_value_success = VerQueryValueW(
-      file_version_info,
-      L"\\VarFileInfo\\Translation",
-      (void**) &lang_buffer,
-      &lang_buffer_size
-  );
-
-  if (!is_ver_query_value_success) {
-    ExitOnWindowsFunctionFailureWithLastError(
-        L"VerQueryValueW",
-        GetLastError()
-    );
-  }
-
-  /* Format text into the file string sub block. */
-  file_string_sub_block_capacity = string_name_len + 8;
-  file_string_sub_block = NULL;
-
-  do {
-    file_string_sub_block_capacity *= 2;
-    file_string_sub_block = realloc(
-        file_string_sub_block,
-        file_string_sub_block_capacity * sizeof(file_string_sub_block[0])
-    );
-
-    swprintf_result = swprintf(
-        file_string_sub_block,
-        file_string_sub_block_capacity,
-        L"\\StringFileInfo\\%04x%04x\\%ls",
-        lang_buffer[0].wLanguage,
-        lang_buffer[0].wCodePage,
-        string_name
-    );
-  } while (swprintf_result == -1);
-
-  /* Query the file string value. */
-  is_ver_query_value_success = VerQueryValueW(
-      file_version_info,
-      file_string_sub_block,
-      (void**) &temp_file_string_value,
-      &file_string_value_size
-  );
-
-  if (!is_ver_query_value_success) {
-    ExitOnWindowsFunctionFailureWithLastError(
-        L"VerQueryValueW",
-        GetLastError()
-    );
-  }
-
-  file_string_value = malloc(file_string_value_size);
-  wcscpy(file_string_value, temp_file_string_value);
-
-free_file_version_info:
-  free(file_version_info);
-
-  return file_string_value;
-}
 
 void InitGameVersion(const wchar_t* game_path, size_t game_path_len) {
   const wchar_t* kProductNameStr = L"ProductName";
